requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pyTelegramBotAPI==4.11.0
requests==2.28.2
urllib3==1.26.15
gunicorn==20.1.0
Django==4.2
django-countries==7.5.1
django-allauth==0.54.0
django-bootstrap4==22.3
django-widget-tweaks==1.4.12
telegram~=0.0.1
python-telegram-bot==20.2
django-telegram-bot==0.6.0
django-crispy-forms==2.0
dispatcher==1.0
